chore: fixes dev-ui-gallery sample
