+++ js/testapps/dev-ui-gallery/src/main/tools.ts
@@ -17,7 +17,7 @@
 import { gemini15Flash } from '@genkit-ai/googleai';
 import { z } from 'genkit';
 import { WeatherSchema } from '../common/types';
-import { ai } from '../index.js';
+import { ai } from '../genkit.js';
 
 ai.defineTool(
   {
@@ -58,8 +58,8 @@ const template = `
   Always try to be as efficient as possible, and request tool calls in batches.
 
   {{role "user"}}
-  Help me decide which is a better place to visit today based on the weather. 
-  I want to be outside as much as possible. Here are the cities I am 
+  Help me decide which is a better place to visit today based on the weather.
+  I want to be outside as much as possible. Here are the cities I am
   considering:\n\n{{#each cities}}{{this}}\n{{/each}}`;
 
 const weatherPrompt = ai.definePrompt(
